AE-1780: Multistate params/tests migration and clean-up
